Refactored out the reviews section to its own file
