+++ src/App.jsx
@@ -10,7 +10,7 @@ import AdminDashboard from "./page/admin/adminDashboard";
 import UserPage from "./page/userPage";
 import About from "./page/about";
 import Checkout from "./page/checkout";
-import Course from "./page/course";
+import Course from "./page/course/course";
 
 import {UserContext} from "./userContext";
 import {useState} from "react";
@@ -37,7 +37,7 @@ function App() {
                 <Route path={"/"} element={<Layout/>}>
                     <Route index element={<Index/>}/>
                     <Route path={"search"} element={<Search/>}/>
-                    <Route path={"course/:id"} element={<Course/>}/>
+                    <Route path={"/course/:id"} element={<Course/>}/>
                     <Route path={"userpage/:id"} element={<UserPage/>}/>
                     <Route path={"order-complete"} element={<OrderComplete/>}/>
                     <Route path={"about"} element={<About/>}/>
